api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "alembic>=1.16.5",
    "async-lru>=2.0.5",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.14.2",
    "docker>=7.1.0",
    "oss2>=2.19.1",
    "fastapi>=0.116.1",
    "filelock>=3.20.0",
    "httpx>=0.28.1",
    "json-repair>=0.53.0",
    "markdownify>=1.2.2",
    "mcp>=1.22.0",
    "openai>=1.107.2",
    "playwright>=1.56.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt[crypto]>=2.10.1",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
]
